pyproject.toml
[project]
name = "phone-authorizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
]
